pyproject.toml
[project]
name = "fl-studio-mcp"
version = "0.1.0"
description = "MCP server for FL Studio automation - enabling AI to control FL Studio for music generation"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastmcp>=0.11.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
flapi = [
    "flapi>=0.2.0",
]
pyflp = [
    "pyflp>=2.2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
fl-studio-mcp = "fl_studio_mcp.fl_studio_server:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
